+++ src/Pair.sol
@@ -19,10 +19,10 @@ abstract contract Pair is IPair, UniswapV2ERC20 {
     string private constant RECOVERER_NAME = "Shared::defaultRecoverer";
     bytes4 private constant SELECTOR = bytes4(keccak256("transfer(address,uint256)"));
 
-    uint public constant MINIMUM_LIQUIDITY = 10 ** 3;
-    uint public constant FEE_ACCURACY = 1_000_000; // 100%
-    uint public constant MAX_PLATFORM_FEE = 500_000; //  50%
-    uint public constant MAX_SWAP_FEE = 20_000; //   2%
+    uint256 public constant MINIMUM_LIQUIDITY = 10 ** 3;
+    uint256 public constant FEE_ACCURACY = 1_000_000; // 100%
+    uint256 public constant MAX_PLATFORM_FEE = 500_000; //  50%
+    uint256 public constant MAX_SWAP_FEE = 20_000; //   2%
 
     GenericFactory public immutable factory;
     address public immutable token0;
@@ -39,12 +39,12 @@ abstract contract Pair is IPair, UniswapV2ERC20 {
     uint112 internal reserve1;
     uint32 internal blockTimestampLast;
 
-    uint public swapFee;
-    uint public customSwapFee = type(uint).max;
+    uint256 public swapFee;
+    uint256 public customSwapFee = type(uint256).max;
     bytes32 internal immutable swapFeeName;
 
-    uint public platformFee;
-    uint public customPlatformFee = type(uint).max;
+    uint256 public platformFee;
+    uint256 public customPlatformFee = type(uint256).max;
 
     modifier onlyFactory() {
         require(msg.sender == address(factory), "P: FORBIDDEN");
@@ -70,7 +70,7 @@ abstract contract Pair is IPair, UniswapV2ERC20 {
         _blockTimestampLast = blockTimestampLast;
     }
 
-    function setCustomSwapFee(uint _customSwapFee) external onlyFactory {
+    function setCustomSwapFee(uint256 _customSwapFee) external onlyFactory {
         // we assume the factory won't spam events, so no early check & return
         emit CustomSwapFeeChanged(customSwapFee, _customSwapFee);
         customSwapFee = _customSwapFee;
@@ -78,7 +78,7 @@ abstract contract Pair is IPair, UniswapV2ERC20 {
         updateSwapFee();
     }
 
-    function setCustomPlatformFee(uint _customPlatformFee) external onlyFactory {
+    function setCustomPlatformFee(uint256 _customPlatformFee) external onlyFactory {
         emit CustomPlatformFeeChanged(customPlatformFee, _customPlatformFee);
         customPlatformFee = _customPlatformFee;
 
@@ -86,7 +86,7 @@ abstract contract Pair is IPair, UniswapV2ERC20 {
     }
 
     function updateSwapFee() public {
-        uint _swapFee = customSwapFee != type(uint).max ? customSwapFee : factory.get(swapFeeName).toUint256();
+        uint256 _swapFee = customSwapFee != type(uint256).max ? customSwapFee : factory.get(swapFeeName).toUint256();
         if (_swapFee == swapFee) return;
 
         require(_swapFee <= MAX_SWAP_FEE, "P: INVALID_SWAP_FEE");
@@ -96,8 +96,8 @@ abstract contract Pair is IPair, UniswapV2ERC20 {
     }
 
     function updatePlatformFee() public {
-        uint _platformFee =
-            customPlatformFee != type(uint).max ? customPlatformFee : factory.read(PLATFORM_FEE_NAME).toUint256();
+        uint256 _platformFee =
+            customPlatformFee != type(uint256).max ? customPlatformFee : factory.read(PLATFORM_FEE_NAME).toUint256();
         if (_platformFee == platformFee) return;
 
         require(_platformFee <= MAX_PLATFORM_FEE, "P: INVALID_PLATFORM_FEE");
@@ -112,16 +112,18 @@ abstract contract Pair is IPair, UniswapV2ERC20 {
         require(token != token1, "P: INVALID_TOKEN_TO_RECOVER");
         require(_recoverer != address(0), "P: RECOVERER_ZERO_ADDRESS");
 
-        uint _amountToRecover = ERC20(token).balanceOf(address(this));
+        uint256 _amountToRecover = ERC20(token).balanceOf(address(this));
 
         _safeTransfer(token, _recoverer, _amountToRecover);
     }
 
-    function _safeTransfer(address token, address to, uint value) internal returns (bool) {
+    function _safeTransfer(address token, address to, uint256 value) internal returns (bool) {
         // solhint-disable-next-line avoid-low-level-calls
         (bool success, bytes memory data) = token.call(abi.encodeWithSelector(SELECTOR, to, value));
         return success && (data.length == 0 || abi.decode(data, (bool)));
     }
 
-    function _update(uint aTotalToken0, uint aTotalToken1, uint112 aReserve0, uint112 aReserve1) internal virtual;
+    function _update(uint256 aTotalToken0, uint256 aTotalToken1, uint112 aReserve0, uint112 aReserve1)
+        internal
+        virtual;
 }
