+++ src/asset-management/AaveManager.sol
@@ -12,21 +12,21 @@ import { IPool } from "src/interfaces/aave/IPool.sol";
 import { IAaveProtocolDataProvider } from "src/interfaces/aave/IAaveProtocolDataProvider.sol";
 
 contract AaveManager is IAssetManager, Owned(msg.sender), ReentrancyGuard {
-    using FixedPointMathLib for uint;
+    using FixedPointMathLib for uint256;
 
-    event FundsInvested(IAssetManagedPair pair, IERC20 token, uint shares);
-    event FundsDivested(IAssetManagedPair pair, IERC20 token, uint shares);
+    event FundsInvested(IAssetManagedPair pair, IERC20 token, uint256 shares);
+    event FundsDivested(IAssetManagedPair pair, IERC20 token, uint256 shares);
 
     /// @dev tracks how many aToken each pair+token owns
-    mapping(IAssetManagedPair => mapping(address => uint)) public shares;
+    mapping(IAssetManagedPair => mapping(address => uint256)) public shares;
 
     /// @dev for each aToken, tracks the total number of shares issued
-    mapping(address => uint) public totalShares;
+    mapping(address => uint256) public totalShares;
 
     /// @dev percentage of the pool's assets, above and below which
     /// the manager will divest the shortfall and invest the excess
-    uint public upperThreshold = 70;
-    uint public lowerThreshold = 30;
+    uint256 public upperThreshold = 70;
+    uint256 public lowerThreshold = 30;
 
     /// @dev this contract itself is immutable and is the source of truth for all relevant addresses for aave
     IPoolAddressesProvider public immutable addressesProvider;
@@ -55,7 +55,7 @@ contract AaveManager is IAssetManager, Owned(msg.sender), ReentrancyGuard {
 
     function _getBalance(IAssetManagedPair aOwner, address aToken) private view returns (uint112 rTokenBalance) {
         address lAaveToken = _getATokenAddress(aToken);
-        uint lTotalShares = totalShares[lAaveToken];
+        uint256 lTotalShares = totalShares[lAaveToken];
         if (lTotalShares == 0) {
             return 0;
         }
@@ -68,12 +68,18 @@ contract AaveManager is IAssetManager, Owned(msg.sender), ReentrancyGuard {
     //////////////////////////////////////////////////////////////////////////*/
 
     /// @notice if token0 or token1 does not have a market in AAVE, the tokens will not be transferred
-    function adjustManagement(IAssetManagedPair aPair, int aAmount0Change, int aAmount1Change) external onlyOwner {
+    function adjustManagement(IAssetManagedPair aPair, int256 aAmount0Change, int256 aAmount1Change)
+        external
+        onlyOwner
+    {
         _adjustManagement(aPair, aAmount0Change, aAmount1Change);
     }
 
-    function _adjustManagement(IAssetManagedPair aPair, int aAmount0Change, int aAmount1Change) private nonReentrant {
-        require(aAmount0Change != type(int).min && aAmount1Change != type(int).min, "AM: CAST_WOULD_OVERFLOW");
+    function _adjustManagement(IAssetManagedPair aPair, int256 aAmount0Change, int256 aAmount1Change)
+        private
+        nonReentrant
+    {
+        require(aAmount0Change != type(int256).min && aAmount1Change != type(int256).min, "AM: CAST_WOULD_OVERFLOW");
 
         IERC20 lToken0 = IERC20(aPair.token0());
         IERC20 lToken1 = IERC20(aPair.token1());
@@ -91,10 +97,10 @@ contract AaveManager is IAssetManager, Owned(msg.sender), ReentrancyGuard {
 
         // withdraw from the market
         if (aAmount0Change < 0) {
-            _doDivest(aPair, lToken0, lToken0AToken, uint(-aAmount0Change));
+            _doDivest(aPair, lToken0, lToken0AToken, uint256(-aAmount0Change));
         }
         if (aAmount1Change < 0) {
-            _doDivest(aPair, lToken1, lToken1AToken, uint(-aAmount1Change));
+            _doDivest(aPair, lToken1, lToken1AToken, uint256(-aAmount1Change));
         }
 
         // transfer tokens to/from the pair
@@ -102,35 +108,35 @@ contract AaveManager is IAssetManager, Owned(msg.sender), ReentrancyGuard {
 
         // transfer the managed tokens to the destination
         if (aAmount0Change > 0) {
-            _doInvest(aPair, lToken0, lToken0AToken, uint(aAmount0Change));
+            _doInvest(aPair, lToken0, lToken0AToken, uint256(aAmount0Change));
         }
         if (aAmount1Change > 0) {
-            _doInvest(aPair, lToken1, lToken1AToken, uint(aAmount1Change));
+            _doInvest(aPair, lToken1, lToken1AToken, uint256(aAmount1Change));
         }
     }
 
-    function _doDivest(IAssetManagedPair aPair, IERC20 aToken, address aAaveToken, uint aAmount) private {
-        uint lShares = _decreaseShares(aPair, address(aToken), aAaveToken, aAmount);
+    function _doDivest(IAssetManagedPair aPair, IERC20 aToken, address aAaveToken, uint256 aAmount) private {
+        uint256 lShares = _decreaseShares(aPair, address(aToken), aAaveToken, aAmount);
         pool.withdraw(address(aToken), aAmount, address(this));
         emit FundsDivested(aPair, aToken, lShares);
         aToken.approve(address(aPair), aAmount);
     }
 
-    function _doInvest(IAssetManagedPair aPair, IERC20 aToken, address aAaveToken, uint aAmount) private {
+    function _doInvest(IAssetManagedPair aPair, IERC20 aToken, address aAaveToken, uint256 aAmount) private {
         require(aToken.balanceOf(address(this)) == aAmount, "AM: TOKEN_AMOUNT_MISMATCH");
-        uint lShares = _increaseShares(aPair, address(aToken), aAaveToken, aAmount);
+        uint256 lShares = _increaseShares(aPair, address(aToken), aAaveToken, aAmount);
         aToken.approve(address(pool), aAmount);
 
         pool.supply(address(aToken), aAmount, address(this), 0);
         emit FundsInvested(aPair, aToken, lShares);
     }
 
-    function setUpperThreshold(uint aUpperThreshold) external onlyOwner {
+    function setUpperThreshold(uint256 aUpperThreshold) external onlyOwner {
         require(aUpperThreshold <= 100 && aUpperThreshold > lowerThreshold, "AM: INVALID_THRESHOLD");
         upperThreshold = aUpperThreshold;
     }
 
-    function setLowerThreshold(uint aLowerThreshold) external onlyOwner {
+    function setLowerThreshold(uint256 aLowerThreshold) external onlyOwner {
         require(aLowerThreshold <= 100 && aLowerThreshold < upperThreshold, "AM: INVALID_THRESHOLD");
         lowerThreshold = aLowerThreshold;
     }
@@ -148,28 +154,28 @@ contract AaveManager is IAssetManager, Owned(msg.sender), ReentrancyGuard {
         uint112 lToken0Managed = _getBalance(lPair, lToken0);
         uint112 lToken1Managed = _getBalance(lPair, lToken1);
 
-        int lAmount0Change = _calculateChangeAmount(lReserve0, lToken0Managed);
-        int lAmount1Change = _calculateChangeAmount(lReserve1, lToken1Managed);
+        int256 lAmount0Change = _calculateChangeAmount(lReserve0, lToken0Managed);
+        int256 lAmount1Change = _calculateChangeAmount(lReserve1, lToken1Managed);
 
         _adjustManagement(lPair, lAmount0Change, lAmount1Change);
     }
 
-    function returnAsset(bool aToken0, uint aAmount) external {
+    function returnAsset(bool aToken0, uint256 aAmount) external {
         require(aAmount > 0, "AM: ZERO_AMOUNT_REQUESTED");
         IAssetManagedPair lPair = IAssetManagedPair(msg.sender);
-        int lAmount0Change = -int(aToken0 ? aAmount : 0);
-        int lAmount1Change = -int(aToken0 ? 0 : aAmount);
+        int256 lAmount0Change = -int256(aToken0 ? aAmount : 0);
+        int256 lAmount1Change = -int256(aToken0 ? 0 : aAmount);
         assert(lAmount0Change < 0 || lAmount1Change < 0);
         _adjustManagement(lPair, lAmount0Change, lAmount1Change);
     }
 
-    function _calculateChangeAmount(uint aReserve, uint aManaged) internal view returns (int rAmountChange) {
-        uint lRatio = aManaged * 100 / aReserve;
+    function _calculateChangeAmount(uint256 aReserve, uint256 aManaged) internal view returns (int256 rAmountChange) {
+        uint256 lRatio = aManaged * 100 / aReserve;
         if (lRatio < lowerThreshold) {
-            rAmountChange = int(aReserve * ((lowerThreshold + upperThreshold) / 2) / 100 - aManaged);
+            rAmountChange = int256(aReserve * ((lowerThreshold + upperThreshold) / 2) / 100 - aManaged);
             assert(rAmountChange > 0);
         } else if (lRatio > upperThreshold) {
-            rAmountChange = int(aReserve * ((lowerThreshold + upperThreshold) / 2) / 100) - int(aManaged);
+            rAmountChange = int256(aReserve * ((lowerThreshold + upperThreshold) / 2) / 100) - int256(aManaged);
             assert(rAmountChange < 0);
         }
     }
@@ -179,26 +185,26 @@ contract AaveManager is IAssetManager, Owned(msg.sender), ReentrancyGuard {
     //////////////////////////////////////////////////////////////////////////*/
 
     /// @dev expresses the exchange rate in terms of how many aTokens per share, scaled by 1e18
-    function _getExchangeRate(address aAaveToken) private view returns (uint rExchangeRate) {
-        uint lTotalShares = totalShares[aAaveToken];
+    function _getExchangeRate(address aAaveToken) private view returns (uint256 rExchangeRate) {
+        uint256 lTotalShares = totalShares[aAaveToken];
         if (lTotalShares == 0) {
             return 1e18;
         }
         rExchangeRate = IERC20(aAaveToken).balanceOf(address(this)).divWadDown(totalShares[aAaveToken]);
     }
 
-    function _increaseShares(IAssetManagedPair aPair, address aToken, address aAaveToken, uint aAmount)
+    function _increaseShares(IAssetManagedPair aPair, address aToken, address aAaveToken, uint256 aAmount)
         private
-        returns (uint rShares)
+        returns (uint256 rShares)
     {
         rShares = aAmount.divWadDown(_getExchangeRate(aAaveToken));
         shares[aPair][aToken] += rShares;
         totalShares[aAaveToken] += rShares;
     }
 
-    function _decreaseShares(IAssetManagedPair aPair, address aToken, address aAaveToken, uint aAmount)
+    function _decreaseShares(IAssetManagedPair aPair, address aToken, address aAaveToken, uint256 aAmount)
         private
-        returns (uint rShares)
+        returns (uint256 rShares)
     {
         rShares = aAmount.divWadDown(_getExchangeRate(aAaveToken));
         shares[aPair][aToken] -= rShares;
