+++ src/asset-management/AssetManagedPair.sol
@@ -36,12 +36,12 @@ abstract contract AssetManagedPair is Pair, IAssetManagedPair {
     uint112 public token0Managed;
     uint112 public token1Managed;
 
-    function _totalToken0() internal view returns (uint) {
-        return IERC20(token0).balanceOf(address(this)) + uint(token0Managed);
+    function _totalToken0() internal view returns (uint256) {
+        return IERC20(token0).balanceOf(address(this)) + uint256(token0Managed);
     }
 
-    function _totalToken1() internal view returns (uint) {
-        return IERC20(token1).balanceOf(address(this)) + uint(token1Managed);
+    function _totalToken1() internal view returns (uint256) {
+        return IERC20(token1).balanceOf(address(this)) + uint256(token1Managed);
     }
 
     function _handleReport(address token, uint112 prevBalance, uint112 newBalance) internal {
@@ -86,15 +86,15 @@ abstract contract AssetManagedPair is Pair, IAssetManagedPair {
         try assetManager.afterLiquidityEvent() { } catch { } // solhint-disable-line no-empty-blocks
     }
 
-    function adjustManagement(int token0Change, int token1Change) external onlyManager {
-        require(token0Change != type(int).min && token1Change != type(int).min, "AMP: CAST_WOULD_OVERFLOW");
+    function adjustManagement(int256 token0Change, int256 token1Change) external onlyManager {
+        require(token0Change != type(int256).min && token1Change != type(int256).min, "AMP: CAST_WOULD_OVERFLOW");
 
         if (token0Change > 0) {
-            uint112 lDelta = uint112(uint(int(token0Change)));
+            uint112 lDelta = uint112(uint256(int256(token0Change)));
             token0Managed += lDelta;
             IERC20(token0).transfer(address(assetManager), lDelta);
         } else if (token0Change < 0) {
-            uint112 lDelta = uint112(uint(int(-token0Change)));
+            uint112 lDelta = uint112(uint256(int256(-token0Change)));
 
             // solhint-disable-next-line reentrancy
             token0Managed -= lDelta;
@@ -103,14 +103,14 @@ abstract contract AssetManagedPair is Pair, IAssetManagedPair {
         }
 
         if (token1Change > 0) {
-            uint112 lDelta = uint112(uint(int(token1Change)));
+            uint112 lDelta = uint112(uint256(int256(token1Change)));
 
             // solhint-disable-next-line reentrancy
             token1Managed += lDelta;
 
             IERC20(token1).transfer(address(assetManager), lDelta);
         } else if (token1Change < 0) {
-            uint112 lDelta = uint112(uint(int(-token1Change)));
+            uint112 lDelta = uint112(uint256(int256(-token1Change)));
 
             // solhint-disable-next-line reentrancy
             token1Managed -= lDelta;
