+++ src/interfaces/IAssetManagedPair.sol
@@ -7,7 +7,7 @@ interface IAssetManagedPair is IPair {
     function token0Managed() external returns (uint112);
     function token1Managed() external returns (uint112);
 
-    function adjustManagement(int token0Change, int token1Change) external;
+    function adjustManagement(int256 token0Change, int256 token1Change) external;
     function assetManager() external returns (IAssetManager);
     function setManager(IAssetManager manager) external;
