+++ src/interfaces/IAssetManager.sol
@@ -6,5 +6,5 @@ import { IAssetManagedPair } from "src/interfaces/IAssetManagedPair.sol";
 interface IAssetManager {
     function getBalance(IAssetManagedPair owner, address token) external returns (uint112 tokenBalance);
     function afterLiquidityEvent() external;
-    function returnAsset(bool aToken0, uint aAmount) external;
+    function returnAsset(bool aToken0, uint256 aAmount) external;
 }
