+++ src/interfaces/IGenericFactory.sol
@@ -5,9 +5,9 @@ interface IGenericFactory {
     function get(bytes32 key) external view returns (bytes32 value);
     function set(bytes32 key, bytes32 value) external;
 
-    function addCurve(bytes calldata initCode) external returns (uint curveId);
+    function addCurve(bytes calldata initCode) external returns (uint256 curveId);
 
     function allPairs() external view returns (address[] memory);
-    function getPair(address tokenA, address tokenB, uint curveId) external view returns (address);
-    function createPair(address tokenA, address tokenB, uint curveId) external returns (address);
+    function getPair(address tokenA, address tokenB, uint256 curveId) external view returns (address);
+    function createPair(address tokenA, address tokenB, uint256 curveId) external returns (address);
 }
