+++ src/interfaces/IPair.sol
@@ -5,13 +5,13 @@ import { IUniswapV2ERC20 } from "src/interfaces/IUniswapV2ERC20.sol";
 
 interface IPair is IUniswapV2ERC20 {
     // solhint-disable-next-line func-name-mixedcase
-    function MINIMUM_LIQUIDITY() external pure returns (uint);
+    function MINIMUM_LIQUIDITY() external pure returns (uint256);
     // solhint-disable-next-line func-name-mixedcase
-    function FEE_ACCURACY() external pure returns (uint);
+    function FEE_ACCURACY() external pure returns (uint256);
     // solhint-disable-next-line func-name-mixedcase
-    function MAX_PLATFORM_FEE() external pure returns (uint);
+    function MAX_PLATFORM_FEE() external pure returns (uint256);
     // solhint-disable-next-line func-name-mixedcase
-    function MAX_SWAP_FEE() external pure returns (uint);
+    function MAX_SWAP_FEE() external pure returns (uint256);
 
     function factory() external returns (GenericFactory);
     function token0() external view returns (address);
@@ -20,8 +20,8 @@ interface IPair is IUniswapV2ERC20 {
     function skim(address to) external;
     function sync() external;
 
-    function mint(address to) external returns (uint liquidity);
-    function burn(address to) external returns (uint amount0, uint amount1);
+    function mint(address to) external returns (uint256 liquidity);
+    function burn(address to) external returns (uint256 amount0, uint256 amount1);
     /**
      * @notice Swaps one token for another. The router must prefund this contract and ensure there isn't too much
      * slippage.
@@ -30,25 +30,25 @@ interface IPair is IUniswapV2ERC20 {
      * @param to address to send the output token and leftover input tokens, callee for the flash swap
      * @param data calls to with this data, in the event of a flash swap
      */
-    function swap(int amount, bool inOrOut, address to, bytes calldata data) external returns (uint amountOut);
+    function swap(int256 amount, bool inOrOut, address to, bytes calldata data) external returns (uint256 amountOut);
 
-    function swapFee() external view returns (uint);
-    function platformFee() external view returns (uint);
-    function customSwapFee() external view returns (uint);
-    function customPlatformFee() external view returns (uint);
-    function setCustomSwapFee(uint _customSwapFee) external;
-    function setCustomPlatformFee(uint _customPlatformFee) external;
+    function swapFee() external view returns (uint256);
+    function platformFee() external view returns (uint256);
+    function customSwapFee() external view returns (uint256);
+    function customPlatformFee() external view returns (uint256);
+    function setCustomSwapFee(uint256 _customSwapFee) external;
+    function setCustomPlatformFee(uint256 _customPlatformFee) external;
     function updateSwapFee() external;
     function updatePlatformFee() external;
 
     function recoverToken(address token) external;
 
-    event Mint(address indexed sender, uint amount0, uint amount1);
-    event Burn(address indexed sender, uint amount0, uint amount1);
-    event Swap(address indexed sender, bool zeroForOne, uint amountIn, uint amountOut, address indexed to);
+    event Mint(address indexed sender, uint256 amount0, uint256 amount1);
+    event Burn(address indexed sender, uint256 amount0, uint256 amount1);
+    event Swap(address indexed sender, bool zeroForOne, uint256 amountIn, uint256 amountOut, address indexed to);
     event Sync(uint112 reserve0, uint112 reserve1);
-    event SwapFeeChanged(uint oldSwapFee, uint newSwapFee);
-    event CustomSwapFeeChanged(uint oldCustomSwapFee, uint newCustomSwapFee);
-    event PlatformFeeChanged(uint oldPlatformFee, uint newPlatformFee);
-    event CustomPlatformFeeChanged(uint oldCustomPlatformFee, uint newCustomPlatformFee);
+    event SwapFeeChanged(uint256 oldSwapFee, uint256 newSwapFee);
+    event CustomSwapFeeChanged(uint256 oldCustomSwapFee, uint256 newCustomSwapFee);
+    event PlatformFeeChanged(uint256 oldPlatformFee, uint256 newPlatformFee);
+    event CustomPlatformFeeChanged(uint256 oldCustomPlatformFee, uint256 newCustomPlatformFee);
 }
